multi_agent_workflow_demo_in_langgraph/requirements.txt
# ---------------- CORE ----------------
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
matplotlib>=3.8.0

# ---------------- LANGCHAIN STACK ----------------
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-experimental>=0.0.50
langgraph>=0.1.0
langsmith>=0.1.0  # optional (tracking)

# ---------------- TOOLS ----------------
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
ipython>=8.12.0

# ---------------- TYPES ----------------
typing-extensions>=4.9.0
pydantic>=2.6.0